feat: Enhance make.py with automatic styles.txt generation
This commit improves the functionality of `make.py` by automatically locating
and modifying the content of the "PROMPT" node based on example prompts defined
in the txt configuration files.

Additionally, the script now generates a "styles.txt" file containing a list of
workflow styles, which is typically included within the release zip for info
purposes. Previously, this file had to be created manually, but with this update,
it is automatically generated alongside other components of the workflow.